04-security sign in register

+++ develop/src/main/java/cz/cvut/fel/nss/vestana/repo/CustomerRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepo extends CrudRepository<Customer, Integer> {
+public interface CustomerRepo extends CrudRepository<Customer, Long> {
 }
