+++ develop/src/main/java/cz/cvut/fel/nss/vestana/repo/ReservationRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservationRepo extends CrudRepository<Reservation, Integer> {
+public interface ReservationRepo extends CrudRepository<Reservation, Long> {
 }
